Grpc timestamps (#1199)
* started processTimestamps() function

* Implemented requestReceived_ and unit tests for getPopulatedSlots

* Implemented timestamps and unit tests for getParam

* Implemented timestamps and unit tests for all methods

* Fixed bubgbot issues

* Modified request-start validation

* Fixed typo

* Changed timestamp format

* Changed timestamp validation and testing to match REST

* Addec valid timestamp test cases

* Updated copyrights

* Fixed requestStart/Received type from double to long

* Improved gRPC timestamp testing

* Added newline at EOF

+++ unittests/cpp/REST/tests/SocketReader_test.cpp
@@ -122,7 +122,7 @@ class RESTSocketReaderTests : public testing::Test, public RESTTest {
         EXPECT_EQ(socketReader.jsonBody(), jsonBody);
         EXPECT_EQ(socketReader.authorizationEnabled(), authz);
         EXPECT_EQ(socketReader.stream(), stream);
-        EXPECT_EQ(socketReader.requestStart(), requestStart != "" ? stod(requestStart) : DEFAULT_REQUEST_START);
+        EXPECT_EQ(socketReader.requestStart(), requestStart != "" ? std::stol(requestStart) : DEFAULT_REQUEST_START);
         EXPECT_TRUE(socketReader.requestReceived() > 0);
     }
